Run py.test with its verbose flag in CI.
Sometimes CI tests hang and then get killed by Jenkins. By default,
py.test just prints dots for each test, which is nice when you are
running it interactively, but when this happens in CI it means we
can't tell which test hangs.

This commit adjusts py.test such that we use it's verbose flag
(-v), which prints the name of each test as it goes. This will be
helpful in identifying which test is getting stuck.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>