Build the integration test containers in parallel.
I recently altered our Jenkins job such that it would not run the
integration tests in parallel. This was to avoid putting so much
load on our Duffy nodes such as to cause the tests to timeout.
However, I also made it so that it does not build the containers
used by the integration tests in parallel anymore.

This commit adjusts the Jenkins job to again run the builds in
parallel.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>