If an integration test fails, keep going.
We recently worked around a CI load issue by running the
integration tests in serial rather than in parallel. However, we
introduced a regression in that a failure in one release causes
Jenkins to halt, rather than proceeding to test the other releases.

This commit makes it such that the CI will proceed even if one
fails so we get all of the results.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>