Use a regex when asserting CLI koji calls.
The CLI download command tests did not work well for multi-build
updates because it asserted builds in the wrong order. This commit
adjusts the test so that it uses a regular expression to assert
that the right builds were downloaded so that the order of the
builds is not being asserted.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>