Raise the CI timeout on bodhi-ci calls to 64 m.
We recently gained a few more CI tests (awesome!) in Bodhi, but now
Bodhi's tests don't always finish before 32 minutes on some Duffy
nodes. Most systems are able to run the tests in less than 32
minutes, but we do occasionally get a node that seems to be too
slow and it hits timeouts. Let's double the timeout.

The better long term fix is to improve Bodhi's performance (though
a lot of the time is also spent waiting on Greenwave which is
beyond our scope to fix). This would make the CI tests go faster
and also improve production performance.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>