CI: Retry on build failures.
CI build failures are often due to network issues and not
necessarily due to problems in Bodhi. This commits adjusts the CI
system to try again on build failures up to 4 times with a 60
second sleep in between. This should make our CI a little more
reliable. The con is that real build failures will take longer to
be reported since they will be tried many times.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>