Add a RateLimiter log filter and document it.
In order to send crash e-mails to admins with Bodhi 4, we need a
way to rate limit the e-mails so we don't send too many repeat
messages about the same problem. This commit introduces
bodhi.server.logging.RateLimiter, which can be used with Python's
logging system to filter messages via the SMTPHandler.

Unfortunately, it is not possible to use logging filters with
Python's configparser logging configs, which is what Pyramid uses
for logging configuration. Due to this, this commit also adds
documentation describing how deployments can use pyramid_sawing to
move the logging configuration into a YAML file. It also documents
the RateLimiter filter. This allows deployments to use the log
filter with Python's built-in SMTPHandler if they don't mind using
a plugin.

This commit also adjusts all places that were configuring logging
to use a new bodhi.server.logging.setup() function that is able to
use the pyramid_sawing YAML file if it is configured, or is able
to use the regular Pyramid INI configuration if the plugin is not
being used. This way there is one way to configure logging in
Bodhi across all processes.

fixes #969

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>