Give bodhi-ci diff_cover a -Z flag.
We recently started sharing the .git folder into the container with
a bind mount for the diff_cover container. We used the Z flag on
the docker bind mount so that the container would have the right
SELinux labeling in the container. Unfortunately, the Z flag also
caused the container to fail to be able to read the .git folder in
the Vagrant guest.

This commit adds a -Z flag so that the tests in CentOS CI can run
with the flag, but running in the Vagrant guest will not use it.
This way the tests work in both environments.

fixes #3166

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>
(cherry picked from commit be4b874a87180079a2ec55d32b90c253a62a9ebf)