Add an index to comments.update_id.
Many queries will pull in update comments, and it was discovered
that the comments table's update_id column was not indexed. This
means that the comments table was being scanned repeatedly when
updates were being retrieved from the database.

This commit lowered the time to query for all updates from a given
user from 5638.934 ms to 14.911 ms in the Vagrant environment,
just 0.264% of the previous time to run the query.

re #3062

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>