z instead of Z on the .git folder for diff-cover
This commit adjusts our Docker bind mount to use the z flag instead
of the Z flag. It is possible that multiple containers run at the
same time with the same content, so the shared label is more
sensible.

This is an attempt to fix an occasional crash in CI with
this traceback:

Traceback (most recent call last):
  File "/usr/bin/diff-cover", line 11, in <module>
    load_entry_point('diff-cover==1.0.7', 'console_scripts', 'diff-cover')()
  File "/usr/lib/python3.7/site-packages/diff_cover/tool.py", line 340, in main
    GitPathTool.set_cwd(directory)
  File "/usr/lib/python3.7/site-packages/diff_cover/git_path.py", line 30, in set_cwd
    cls._root = cls._git_root()
  File "/usr/lib/python3.7/site-packages/diff_cover/git_path.py", line 64, in _git_root
    git_root = execute(command)[0]
  File "/usr/lib/python3.7/site-packages/diff_cover/command_runner.py", line 44, in execute
    raise CommandError(stderr)
diff_cover.command_runner.CommandError: fatal: not a git repository (or any of the parent directories): .git

I suspect it may be that this happens because another container
already has the .git folder bind mounted.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>