Use greenwave subject_identifier to store missing tests in the UI.
This commmit makes use of the subject_identifier (build nvr) in
greenwave response to store the related missing tests.
The missing tests are then displayed in the list of automated test on
the UI.

the current requirement.item is a JS object with key values pair inside,
the code that build the row to display the tests results is using the build
NVR to access the data from the missing_test object.
In the current response from greenwave we have the build NVR in
requirement.subject_identifier hence the change.

My guess is that either this code never worked or that greenwave used to return
a build NVR in requirement.item and that has changed.

Signed-off-by: Clement Verna <cverna@tutanota.com>