Make the consumer optionally import the composer.
The composer has a significant dependency tree and we do not want
to have to install that tree for processes that won't be handling
composer.start messages. The downstream RPMs of Bodhi have split
the composer into its own RPM, so in the spirit of supporting that
effort this commit adjusts the consumer such that it conditionally
imports the composer. If the composer is not installed and it
receives a composer.start message, it will raise a Nack and log an
error.

fixes #3231

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>