Do not crash the updates handler upon messages.
The updates handler was crashing in staging because it was
expecting the old fedmsg but was receiving a fedora-messaging
Message. This commit adjusts the handler to use Bodhi's new
Message objects. It also adds a new field to the schema of one
since the updates handler was using that field.

fixes #3233

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>