Handle Exceptions in message consumption
If any Exceptions occur while handling received messages, we
previously would just crash our message handling process. This was
mostly OK because systemd would just restart it, but the problem
is that our RateLimiter log filter would get its state reset,
which resulted in more error e-mails getting sent than desirable.

With this commit we now log and Nack any message if an Exception
is raised in the handlers.

fixes #3238

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>