Make sure %{uid} in krb ccache gets replaced with the effective UID
This makes it so runs by different users get evaluated to different
ccache's.
This prevents issues when a process is run by another user, and then
the normal procedures would be unable to authenticate to buildsys
until the incorrectly-owned ccache is removed.

For example: bodhi composes might fail to log in to koji after an
admin used bodhi-shell to fix something else, because the bodhi-shell
might have run as root.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>