Remove duplicated logic from the update template
The update template had some duplicated boolean logic with
Update.install_command(). This commit removes that logic by
adjusting install_command() such that it returns the empty string
when it cannot generate an install command, rather than raising an
Exception. This way the template can just run the method and only
display the install command if it is not the empty string.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>