Correct a message type causing missing coverage
I noticed today that we were missing test coverage on one line and
one jump in the updates handler. It turned out that when I
converted the codebase to use Fedora Messages, I had accidentally
chosen the incorrect type for one of the messages and this resulted
in missing coverage on a line and a jump.

This commit corrects that error and raises coverage again to 100%
in the updates handler.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>