Organize all tests for meets_testing_requirements
The tests for meets_testing_requirements() were getting very long
names and weren't cleanly grouped together. This commit gives them
their own test class and simpler names so they are easier to
locate and modify. This exercise also made it clear that we have
poor test case coverage on the function, with only three tests
that don't deal with test gating.

This will ease development of #3282.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>