meets_testing_requirements() complies with policy
There were two ways in which Bodhi's
meets_testing_requirements() did not comply with Fedora's update
policy[0]. For one, updates (both critpath and regular) are
supposed to be able to go stable when they reach two karma (the
critpath.min_karma setting). Secondly, regular updates don't need
to be set to autokarma to be allowed to push based on karma.

This commit fixes both of those scenarios and adds several unit
tests to ensure correct operation of the function, as there were
only three existing tests that did not check these situations or
others.

It also refactors some repeated code to apply the DRY principle so
we can share the code to calculate a release's critpath.min_karma.

As a result, the "Push to Stable" button should now appear at the
correct times based on the update policy.

fixes #1307
fixes #1508
fixes #1796
fixes #3282

[0] https://docs.fedoraproject.org/en-US/fesco/Updates_Policy/#updates-to-critical-path-packages

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>