Don't raise Exception on non-existing composes.
Now that we are using Fedora messaging, we need to be careful not
to raise Exceptions on permanent failures. One example of a
permanent failure happened this weekend, where some messages in
Bodhi's composer queue were referencing Composes that no longer
exist. When sqlalchemy raised an Exception upon failure to
retrieve these Composes, we were Nacking the message, which put it
back into the queue, which caused us to Nack loop eternally.

This commit catches that particular Exception and causes the
Composer to no-op when receiving such messages so that they get
Acked.

fixes #3318

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>