Fix downloading packages for updates with multiple builds
I inadvertently broke this in #3148 - sorry. I tried to make
the handling of the args list cleaner in that commit, but
overlooked that initializing it outside of the `for build in`
loop meant it'd just keep getting appended to, it doesn't get
recreated for each build - so by the time you get to the second
build in an update, you're trying to do something like:

koji download-build foo-1.0-1 bar-1.0-1

which isn't allowed (that's why we run one command per build in
the first place, life would be easier if Koji let us pass it
multiple NVRs...)

Fix this simply by (re)initializing `args` inside the loop, not
outside it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>
Signed-off-by: Nils Philippsen <nils@redhat.com>