Make sure we can set a release mandatory days in testing to 0.
In the case of the rawhide release we need to set the mandatory days
in testing to 0. This will be achieve by the following setting
f31.pre_beta.mandatory_days_in_testing = 0. This commit ensures that
we the logic will return 0.

Signed-off-by: Clement Verna <cverna@tutanota.com>

Convert the mandatory_days_in_testing in integer only if not None

Signed-off-by: Clement Verna <cverna@tutanota.com>