Fix skopeo tests
Prior to pytest 5.0, ExceptionInfo objects had a __str__() method of
their own which rendered the wrapped exception as a string. It was
removed in version 5.0 which made calling str() on the object use
__repr__() as a fallback which gives totally different results.

Here we use str() on the .value member (i.e. the wrapped exception) and
a variable name that makes it (more) explicit that the context manager
isn't an exception.

Signed-off-by: Michal Konečný <mkonecny@redhat.com>
Signed-off-by: Nils Philippsen <nils@redhat.com>
(cherry picked from commit adbc20708e349f62d2602b0d2e0bd29e472d8a31)