Automatically create the rawhide updates in a TESTING state
For updates that are automatically created for rawhide there is no
need to have them go through the PENDING state. We can just directly
create them in the TESTING state, let them be tested by CI and let
the approve_testing cron job move them to STABLE once they have
passed their CI.

In this commit we are also dropping the call to move the update into
pending testing, since it is directly in testing, but we are adding
a comment (from "bodhi") about the update being automatically
created so there is an email notification sent about this update.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>