Mark the update as stable if they are not included in compose ran by bodhi
In regular releases we have:
- pending-testing
- testing
- pending-stable
- stable

For the rawhide release in Fedora, we have:
- testing
- stable
Updates that are in testing and whose gating status is "green", are
pushed to stable via the approve_testing script. On "regular"
releases they will go from testing to pending-stable, on "rawhide"
we want them to go from testing to stable directly since the compose
process that normally moves the build from pending-stable to stable
is not run by bodhi in rawhide.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>