Handle previously existing pending updates
If the user manually creates a bodhi update for a release for which the
update should have been automatically created, that update will stay in
a pending state with nothing moving it forward. When we finally receive
the message from Koji about a build being tagged in the tag of interest,
we will discard it since there is already an update for it. This results
in the update effectively being stuck in Bodhi.

With this change, if Bodhi is notified of a build tagged in a Koji tag
of interest, we will check if there is an existing pending update for it
and if so, we will move it to "Testing" and check its gating status,
effectively allowing to move the update to the next stage from where it
can be picked by the regular workflow.

Fixes https://github.com/fedora-infra/bodhi/issues/3414

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>
(cherry picked from commit 1de9c95227eb39481c6b4f50ead617c405b0719a)