Handle orphan builds in automatic updates consumer
In the Bodhi database, builds can exist without an associated update,
e.g. if a user creates a buildroot override with a build that hasn't
been added to an update yet.

That means that we have to check if the build is associated with an
update before checking that update's status.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>
Signed-off-by: Nils Philippsen <nils@redhat.com>
(cherry picked from commit 01a8fe7109cda7eb285a771cd429dbfbda07c917)