mark updates not composed by bodhi as pushed when stable
when an update is composed by bodhi, composer.py marked
the update as pushed. However, in the case of an update
that is not composed_by_bodhi, (e.g an automatically
created rawhide update), the update's pushed boolean would
always remain false. The update.html template uses the
pushed boolean to help decide what action buttons to show,
and consequently, an automatically created rawhide update
that is in stable would still show the push to testing buttons
at the top of this page (#3451)

This commit changes the approve_testing.py script to update the
update's pushed boolean, and the date pushed boolean when an
update that is not composed by bodhi is marked as stable.

Resolves: #3451
Signed-off-by: Ryan Lerch <rlerch@redhat.com>