Delete additional tags once an side tag update was pushed to stable.
This commits makes sure that the additional tags and the original
side tag are deleted once the update was pushed to stable.

Signed-off-by: Nils Philippsen <nils@redhat.com>

Log identical timestamps for the same operation

Signed-off-by: Nils Philippsen <nils@redhat.com>

Remove side and aux tags when closing out updates

Fixes: #3476

Signed-off-by: Nils Philippsen <nils@redhat.com>

Refactor the approve testing script for multi build updates.

This commit makes the approve testing script easier to follow
by splitting the logic for single build updates, multi build updates
and updates composed by bodhi.
Notes: For updates that are not composed by bodhi we do not
use the set_requests method anymore. This is really only useful
for the composer and makes things simplier in the rawhide context.

Signed-off-by: Clement Verna <cverna@tutanota.com>