Check if it is safe to tag a multi build update in stable.
This commit add the logic to check if a multi build update
has builds that conflicts with the stable tag. To do that we
are checking for each package present in the update if a more
recent build exists in the stable tag. If that is the case we
consider that this package build is conflicting and the update
will not be pushed to stable.

If none of the build present in the update are conflicting then
we proceed with pushing the update to stable.

Signed-off-by: Clement Verna <cverna@tutanota.com>