Unify rawhide simple build update with multi build update.
This commits changes how bodhi automatically creates single build
updates for rawhide. Instead of creating the update directly in the
testing status it creates it in pending. The update is then moved in
testing when the build is signed.
Once the update is moved to testing bodhi sends a fedora-messaging
message to trigger the CI tests. Results of these tests will be received
by the greenwave consumer.

Fixes #3513

Signed-off-by: Clement Verna <cverna@tutanota.com>