remove critpath_karma from the UI
for a long time, there has been a critpath_karma field
in an updates comments that users could use on critical path
updates. However, it was not used by bodhi to do anything.

This PR removes the critpath_karma from the ui, but leaves
everything else for now. see #2194 for more information.

Additionally, the bodhi cli never had support for changing
the critpath_karma.

Fixes: #2194
Fixes: #2183

Signed-off-by: Ryan Lerch <rlerch@redhat.com>