Tidy up the UI a little more
fixes a few things in the new UI:

1. Fix updates not loading on user profile page
the user profile page was failing to load the updates
of a user due to a bug introduced in #3511. The user
profile page pulls html down from the server to show
the list. However the template was also trying to add
the javascript for filtering, which was causing a 500
error.

2. make the suggestion on the update page a <p>.
previously, the suggestion (suggesting the user either
reboot or logout to properly test the package) was in
an alert, which kinda stole the show. This changes that
box to a more subtle paragraph.

3. Clean up the composes and releases pages

Signed-off-by: Ryan Lerch <rlerch@redhat.com>