Reload update page when commenting
Previously, when adding a comment to an update, the javascript
dynamically added the comment to the bottom of the list, but did not
update the karma. In this commit, we just reload the entire page rather
than fuss with adding the information correctly after the async comment
form returns a success.

Fixes: #3502
Signed-off-by: Ryan Lerch <rlerch@redhat.com>