Move multi build update that failed to merge in rawhide to pending.
This commit move a multi build rawhide update that was tested but failed
to merge in rawhide (more recent build found in rawhide buildroot)
to pending.
Once in pending the maintainer needs to edit the update to either
add a new build or remove a conflicting build.
On the edit action bodhi checks if all the builds in the update are
signed and moved the update back to testing. If the build are not signed
we wait in pending until we receive the signed message from koji.

Fixes #3514

Signed-off-by: Clement Verna <cverna@tutanota.com>
(cherry picked from commit 9d380c1fc23b42f5d9a1d4307b10760168175d8b)