Move the --user and --password arguments in a group with --openid-api
This way action which allow to set the --openid-api arguments will not
forget to add the --user and --password ones (as did the waive and
request-tests).
This commit also fixes the waive and request-tests action so they allow
logging in from the CLI via the --user and --password arguments as do
other actions requiring to be logged in.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>
(cherry picked from commit 1891d269cb6145685f8c23652cff2fcb0f11020a)