fix overrides list on public user profile page
previously, the user public profile page was not showing the list of
buildroot overrides. This was due to the overrides.html returning a
block that should only be returned when rendering the page as a whole,
rather than the fragment used in this case. This fixes this issue.

Fixes: #3567

Signed-off-by: Ryan Lerch <rlerch@redhat.com>
(cherry picked from commit f29492799e3ee3281c6f01c0ad0d0d0c8ff69197)