Allow autosign on -testing-pending
This commit allows to specify the permissions when creating a tag in
koji for side-tag updates we are creating two other tags -signing-pending
and -testing-pending. The later of the two should allow autosign to move
builds into.
This commit thus allows specifying permissions when creating a tag as
well as give autosign permissions on -testing-pending tag for side-tag
based updates.

Fixes https://github.com/fedora-infra/bodhi/issues/3627

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>