Do not double enquote the update title
Basically, if we specify amp=True on update.beautify_title(), it will
replace the & sign with its html quoted version: &amp;
But since we're encoding the output of this function into html again,
the & in &amp; get encoded again resulting in a "&amp;amp;" string which
is not valid html and is breaking the test suite.

Thanks to Nils for getting the idea of what is going on!

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>