work around slow /latest_candidates calls
Commit 9d6dffe15b96ab52d010aa49253ff676365f0676 introduces hide_existing
to the /latest_candidates API endpoint but setting it true slows down
initial calls in the form (with empty or short prefixes) unduly. This is
because it queries the database for each build returned by koji which
can be on the order of several thousands.

Set hide_existing=false for the time being, until we have a better
performing way to filter out builds which are related to updates
already.

Signed-off-by: Nils Philippsen <nils@redhat.com>