Speed up hide_existing=true when listing candidates
Previously, the database was queried for each build returned from koji,
which could result in thousands of individual queries.

Instead, query the database once for known builds associated with an
update and filter out returned builds found in this set. This set isn't
filtered for active releases because a build might be associated with an
update for an archived release but still inherited into one of the tags
we query Koji for. It performs well enough at the moment but the set can
easily be cached if not.

Additionally, revert the workaround we put in place (commit
7a83faf579bc45fcf524bea87c6afc26d8a628f6) and test the feature also
for builds/tags of an archived release.

Fixes: #3687

Signed-off-by: Nils Philippsen <nils@redhat.com>