make bodhi comments render as markdown
previously, in the old UI, the bodhi user comments rendered
as markdown in the updates view. This was mainly used to provide
clickable links in the bodhi comments other updates. However with the
introduction of the new UI, this regressed and the comments were just
shown as text, not rendered markdown.

This commit changes the layout of bodhi comments so we can present the
rendered markdown properly.

Fixes: #3723

Signed-off-by: Ryan Lerch <rlerch@redhat.com>