Manage single build update conflicting builds.
This commits makes sure that we deal with the single build update
when looking at conflicting build in the rawhide buildroot.
If a single build update contain a build that has a more recent build
in the buildroot the update is marked as obsolete and the build pending
testing tag and candidate tag are removed.

Previously the script was trying to remove the update testing side tag
which does not exists for a single build update.

Fixes #3828

Signed-off-by: Clement Verna <cverna@tutanota.com>
(cherry picked from commit dfa43b9596f5dd0922442eac258cf1f4699e2926)