Ignore celery results we don't use.
It is recommended in the celery documentation to ignore tasks results
that are not used. Currently the only task's results used are for
the tasks triggered by celery-beat, the other task's results then can
be ignored.

In production it was also observed that some tasks hanged and never
returned making the web request hang too, which lead to having database
session left open.

This commit prevents such a scenario since triggering a task with no results
will return immediately.

Signed-off-by: Clement Verna <cverna@tutanota.com>