Do not let update through when bodhi fails to talk to greenwave
This commit removes the greenwave_failed status of
the test_gating_status update attribute. Instead in case of
a failure to communicate to greenwave we set the status to
waiting and rely on the fedora-messaging greenwave consumer
to update the gating status.

The greenwave_failed status is still present in the database
table and old updates having that status are not changed.

Fixes #4089

Signed-off-by: Clement Verna <cverna@tutanota.com>