Query different Greenwave contexts for critpath updates
This tweaks Bodhi to query Greenwave with a different decision
context for critical path updates. This will allow us to require
openQA tests pass for these updates (we cannot require this for
all updates as the tests are not run on non-critpath updates).

We also update the Greenwave config in the CI setup to match the
current state of the real Fedora config, including changes I
have proposed to add the appropriate policy for this new context,
and update the tests.

Signed-off-by: Adam Williamson <awilliam@redhat.com>