Ensure updates with stable_comment are pushed to stable automatically
The `approve_testing` task was set to ignore updates which already had
stable_comment.
This was not correct and prevented some updates from being pushed to
stable when they reached the stable_days threshold.

As an example, consider an update for which the associated release has a
mandatory_days_in testing set to 7 and the update is created with a
stable_days threshold of 10.
After 7 days in testing, the update gets the stable_comment, since the
maintainer can now push it to stable.
The autopush is performed after 3 more days, but this didn't happen
because of the approve_testing task ignoring all updates with
stable_comment.

Signed-off-by: Mattia Verga <mattia.verga@tiscali.it>