Correct 're-trigger' and 'agent' when re-triggering tests (#4217)
Per #4217, it seems like we intend for `re-trigger` to be True
and/or `agent` to be the username when a
koji-build-group.build.complete message is produced by a user
requesting a test re-trigger, but we do not actually do either at
present, `re-trigger` is always False and `agent` is always
"bodhi". This changes both; it can obviously be amended to only
do one or the other if desired.

Signed-off-by: Adam Williamson <awilliam@redhat.com>