Don't set gating status twice when creating update (#4219)
On the normal path of creating an update - where it is created
and immediately submitted to updates-testing - we're setting
the gating status twice. We call `set_request`, which calls
`update_test_gating_status()` if the request is updates-testing
and gating is required, and then almost immediately afterwards,
we straight up set the status to 'waiting'. That doesn't make
much sense.

This changes things so that the `new()` method doesn't hard set
the status to 'waiting' if it's calling `set_request` and the
request is to updates-testing, as is usually the case.

Signed-off-by: Adam Williamson <awilliam@redhat.com>