Add "N" to the version model
Fedora's bodhi just added a "EPEL-8N" release that has
epel8-next-updates and epel8-next-updates-testing. This is to build
against CentOS stream in those cases where normal EPEL packages are no
longer compatible due to stream changes. The bodhi module doesn't
understand the 'N' here and can't figure out the version. If we add it
into the model it's fine. :)

Signed-off-by: Kevin Fenzi <kevin@scrye.com>