Consumers: replace Greenwave with ResultsDB and WaiverDB (#4224)
This replaces the consumer that listens for Greenwave 'decision
change' messages with consumers that listen for ResultsDB and
WaiverDB "new result" / "new waiver" messages and update gating
decisions if the message may cause a change.

The goal here is to make it possible to remove the "decision
change" message code from Greenwave entirely, because in order
to publish those messages, Greenwave has to duplicate a lot of
Bodhi's knowledge about what kinds of decisions to ask for. This
being in Greenwave is really against the intended design of the
system, and means whenever anything changes about what kinds of
decision Bodhi might request, we need to change both Bodhi and
the Greenwave "decision change" publishing code.

Signed-off-by: Adam Williamson <awilliam@redhat.com>